pyproject.toml
[tool.poetry]
name = "imagine-freediving"
version = "0.1.0"
description = ""
authors = ["Djane Rey Mabelin <thedjaney@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
freediving = { git = "git@github.com:djaney/freedive.git", tag="v0.2.0"}
gtts = "^2.5.1"
pydub = "^0.25.1"
moviepy = "^1.0.3"
pillow = "<10.0.0" # resize not working in modern version
click = "^8.1.7"

[tool.poetry.scripts]
freedive = { reference = "imagine_freediving.cli:main", type = "console" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
